Move configuration package into 'uml' as subpackage as it configures the UML being generated.
