+++ src/main/java/nl/talsmasoftware/umldoclet/uml/TypeName.java
@@ -15,7 +15,7 @@
  */
 package nl.talsmasoftware.umldoclet.uml;
 
-import nl.talsmasoftware.umldoclet.configuration.TypeDisplay;
+import nl.talsmasoftware.umldoclet.uml.configuration.TypeDisplay;
 
 import java.io.IOException;
