+++ src/test/resources/default-maven-opts/packages
@@ -1,4 +1,4 @@
-nl.talsmasoftware.umldoclet.configuration
+nl.talsmasoftware.umldoclet.uml.configuration
 nl.talsmasoftware.umldoclet.javadoc
 nl.talsmasoftware.umldoclet.logging
 nl.talsmasoftware.umldoclet.model
